cccc

+++ demo/demo/src/main/kotlin/com/example/demo/domain/ScheduledNetworkTopology.kt
@@ -2,5 +2,5 @@ package com.example.demo.domain
 
 class ScheduledNetworkTopology(
     val adjacency: Map<Sensor, List<Sensor>>,
-    val dutyCycles: Map<Sensor, Int>
+    val dutyCycles: Map<Sensor, Double>
 )
