+++ demo/demo/src/main/kotlin/com/example/demo/domain/Sensor.kt
@@ -2,8 +2,8 @@ package com.example.demo.domain
 
 class Sensor(
     val id: String,
-    val desiredDutyCycle:Int,
-    val tolerance: Int? = null
+    val desiredDutyCycle: Double,
+    val tolerance: Double? = null
 )
 {
     override fun toString(): String = id
