+++ demo/demo/src/main/kotlin/com/example/demo/pipeline/SensorGroupInput.kt
@@ -2,6 +2,6 @@ package com.example.demo.pipeline
 
 data class SensorGroupInput(
     val sensorIds: List<String>,
-    val desiredDutyCycle: Int,
-    val tolerance: Int
+    val desiredDutyCycle: Double,
+    val tolerance: Double
 )
